update on my last pull-request with requested changes from niel
